backend/requirements.txt
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
flask-shell-ipython==0.4.1
SQLAlchemy-Utils==0.36.1
marshmallow==3.3.0
attrs==19.3.0
funcy==1.14
environs==6.1.0
lob==4.0.1
pylint==2.4.4
autopep8==1.4.4
ipython==7.9.0
ipdb==0.12.2
pytest==5.3.1
pytest-flask==0.15.0
factory_boy==2.12.0
